requirements.txt
langchain==0.3.0
langchain-anthropic==0.3.10
langchain-cohere==0.3.1
langchain-community==0.3.0
langchain-core==0.3.63
langchain-text-splitters==0.3.2
faiss-cpu==1.8.0
spotipy==2.24.0
cohere==5.11.0
anthropic==0.84.0
python-dotenv==1.0.1
numpy==1.26.4
pydantic==2.9.2
pypdf==4.3.1
beautifulsoup4==4.14.3
lxml==6.0.2
